2주차 코드리뷰 요청

See merge request ai_track/class_07/data_project/team03/ecobucks!66

+++ front/src/pages/Challenge/ChallengeRead.jsx
@@ -1,30 +1,40 @@
-import React, { useState, useContext } from "react";
+import React, { useState, useContext, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import { Button, Container, Card, Modal } from "react-bootstrap";
+import {
+  Button,
+  Container,
+  Card,
+  OverlayTrigger,
+  Tooltip,
+} from "react-bootstrap";
 import * as Api from "../../api";
 
 import ChallengeParticipate from "./ChallengeParticipate";
 import { UserStateContext } from "../../context/user/UserProvider";
-
 import ChallengeUpdate from "./ChallengeUpdate";
+import ChallengeComments from "./ChallengeComments";
 
 const ChallengeRead = ({ challenge, onBackToListClick }) => {
-  const [showModal, setShowModal] = useState(false);
+  const [showParticipateModal, setShowParticipateModal] = useState(false);
   const [showUpdateModal, setShowUpdateModal] = useState(false);
+  const [comments, setComments] = useState([]);
+  const [editedComment, setEditedComment] = useState("");
 
   const userState = useContext(UserStateContext);
-  const navigate = useNavigate();
 
   const handleJoinClick = () => {
-    setShowModal(true);
+    setShowParticipateModal(true);
   };
 
   const handleUpdateClick = () => {
+    if (challenge.participantsCount >= 1) {
+      return;
+    }
     setShowUpdateModal(true);
   };
 
-  const handleCloseModal = () => {
-    setShowModal(false);
+  const handleCloseParticipateModal = () => {
+    setShowParticipateModal(false);
   };
 
   const handleCloseUpdateModal = () => {
@@ -36,22 +46,40 @@ const ChallengeRead = ({ challenge, onBackToListClick }) => {
     return date.toLocaleDateString(); // Format date as 'YYYY-MM-DD'
   };
 
-  const isCurrentUserAuthor = userState.user._id === challenge.user_id._id;
+  const isCurrentUserAuthor = userState.user._id === challenge.userId._id;
 
   const handleDeleteClick = async () => {
+    if (challenge.participantsCount >= 1) {
+      alert("참가자가 1명 이상이므로 수정, 삭제할 수 없습니다");
+      return;
+    }
     const confirmDelete = window.confirm("챌린지를 삭제하시겠습니까?");
     if (confirmDelete) {
       try {
         await Api.delete(`challenges/${challenge._id}`);
+        window.location.reload();
       } catch (err) {
-        alert("챌린지 삭제에 실패하였습니다.");
+        alert(err.res.data);
       }
     }
   };
-  
+
+  useEffect(() => {
+    const fetchComments = async () => {
+      try {
+        const res = await Api.get(`challenges/${challenge._id}/comments`);
+        console.log("코멘트", res.data);
+        setComments(res.data);
+      } catch (error) {
+        console.log("Error fetching comments:", error);
+      }
+    };
+
+    fetchComments();
+  }, [challenge]);
 
   return (
-    <Container>
+    <Container className="mt-5">
       <h2>챌린지 내용 확인</h2>
       <Card className="m-2">
         <Card.Body>
@@ -62,52 +90,77 @@ const ChallengeRead = ({ challenge, onBackToListClick }) => {
             <br />
             마감일자: {formatDate(challenge.dueDate)}
             <br />
-            작성자: {challenge.user_id._id}
+            작성자: {challenge.userId.username}
             <br />
             참여인원: {challenge.participantsCount.toLocaleString()} 명
           </Card.Text>
+
+          <ChallengeComments challenge={challenge} />
+
         </Card.Body>
       </Card>
-      <Button className="mt-3" onClick={handleJoinClick}>
-        참가하기
-      </Button>
+
+      <>
+        <Button className="mt-3" onClick={handleJoinClick}>
+          참가하기
+        </Button>
+      </>
       {isCurrentUserAuthor && (
         <>
-          <Button className="mt-3" onClick={handleUpdateClick}>
-            수정하기
-          </Button>
-          <Button className="mt-3" onClick={handleDeleteClick}>
-            삭제하기
-          </Button>
+          <OverlayTrigger
+            placement="top"
+            overlay={
+              <Tooltip id="tooltip-disabled">
+                참가자가 있을 시 수정, 삭제할 수 없습니다.
+              </Tooltip>
+            }
+          >
+            <a>
+              <Button
+                className="mt-3"
+                onClick={handleUpdateClick}
+                disabled={challenge.participantsCount >= 1}
+              >
+                수정하기
+              </Button>
+            </a>
+          </OverlayTrigger>
+
+          <OverlayTrigger
+            placement="top"
+            overlay={
+              <Tooltip id="tooltip-disabled">
+                참가자가 있을 시 수정, 삭제할 수 없습니다.
+              </Tooltip>
+            }
+          >
+            <a>
+              <Button
+                className="mt-3"
+                onClick={handleDeleteClick}
+                disabled={challenge.participantsCount >= 1}
+              >
+                삭제하기
+              </Button>
+            </a>
+          </OverlayTrigger>
         </>
       )}
       <Button onClick={onBackToListClick} className="mt-3">
         목록으로
       </Button>
 
+      <ChallengeParticipate
+        show={showParticipateModal}
+        onClose={handleCloseParticipateModal}
+        challenge={challenge}
+      />
 
-
-      <Modal show={showModal} onHide={handleCloseModal}>
-        <Modal.Header closeButton>
-          <Modal.Title>참가하기</Modal.Title>
-        </Modal.Header>
-        <ChallengeParticipate onClose={handleCloseModal} />
-      </Modal>
-
-
-
-      <Modal size='lg' show={showUpdateModal} onHide={handleCloseUpdateModal}>
-        <Modal.Header closeButton>
-          <Modal.Title>챌린지 수정</Modal.Title>
-        </Modal.Header>
-        <ChallengeUpdate
-          challenge={challenge}
-          onClose={handleCloseUpdateModal}
-        />
-      </Modal>
-
-
-
+      <ChallengeUpdate
+        show={showUpdateModal}
+        onClose={handleCloseUpdateModal}
+        challenge={challenge}
+      />
     </Container>
   );
 };
